requirements.txt
anthropic>=0.39.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pypdf>=4.0.0

# Security (Phase 2)
cryptography>=42.0.0
keyring>=25.0.0

# Testing
pytest>=8.0.0
